requirements.txt
openai>=1.12.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
sentence-transformers>=2.2.0
plotly>=5.18.0
python-dotenv>=1.0.0
tqdm>=4.66.0

